Move Elevation to separate module and simplify Elevation API.
